+++ server/chalicelib/s3.py
@@ -36,14 +36,14 @@ def get_live_folder(stop_id: str):
     if is_bus(stop_id):
         return "daily-bus-data"
     else:
-        return "daily-data"
+        return "daily-rapid-data"
 
 
-def download_one_event_file(date, stop_id: str, live=False):
+def download_one_event_file(date, stop_id: str, use_live_data=False):
     """As advertised: single event file from s3"""
     year, month, day = date.year, date.month, date.day
 
-    if live:
+    if use_live_data:
         folder = get_live_folder(stop_id)
         key = f"Events-live/{folder}/{stop_id}/Year={year}/Month={month}/Day={day}/events.csv.gz"
     else:
@@ -58,8 +58,8 @@ def download_one_event_file(date, stop_id: str, live=False):
         if ex.response["Error"]["Code"] == "NoSuchKey":
             # raise Exception(f"Data not available on S3 for key {key} ") from None
             print(f"WARNING: No data available on S3 for key: {key}")
-            if not live and is_bus(stop_id):
-                return download_one_event_file(date, stop_id, live=True)
+            if not use_live_data and is_bus(stop_id):
+                return download_one_event_file(date, stop_id, use_live_data=True)
             return []
         else:
             raise
@@ -87,3 +87,16 @@ def download_events(sdate, edate, stops: list):
     result = parallel_download_events(datestops)
     result = filter(lambda row: sdate.strftime("%Y-%m-%d") <= row["service_date"] <= edate.strftime("%Y-%m-%d"), result)
     return sorted(result, key=lambda row: row["event_time"])
+
+
+def get_all_s3_objects(s3, **base_kwargs):
+    continuation_token = None
+    while True:
+        list_kwargs = dict(MaxKeys=1000, **base_kwargs)
+        if continuation_token:
+            list_kwargs["ContinuationToken"] = continuation_token
+        response = s3.list_objects_v2(**list_kwargs)
+        yield from response.get("Contents", [])
+        if not response.get("IsTruncated"):  # At the end of the list?
+            break
+        continuation_token = response.get("NextContinuationToken")
